Move specific output file parsing to specific parameters
