+++ DICUI.Library/Aaru/Parameters.cs
@@ -3,7 +3,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
+using System.Xml;
+using System.Xml.Schema;
 using DICUI.Data;
+using DICUI.Utilities;
 
 namespace DICUI.Aaru
 {
@@ -1367,6 +1370,37 @@ namespace DICUI.Aaru
         }
 
         /// <summary>
+        /// Generate a SubmissionInfo for the output files
+        /// </summary>
+        /// <param name="info">Base submission info to fill in specifics for</param>
+        /// <param name="basePath">Base filename and path to use for checking</param>
+        /// <param name="system">KnownSystem type representing the media</param>
+        /// <param name="type">MediaType type representing the media</param>
+        /// <param name="drive">Drive representing the disc to get information from</param>
+        /// <returns></returns>
+        public override void GenerateSubmissionInfo(SubmissionInfo info, string basePath, KnownSystem? system, MediaType? type, Drive drive)
+        {
+            // TODO: Fill in submission info specifics for Aaru
+            string outputDirectory = Path.GetDirectoryName(basePath);
+
+            // Fill in the hash data
+            info.TracksAndWriteOffsets.ClrMameProData = GenerateDatfile(basePath + ".cicm.xml");
+
+            switch (type)
+            {
+                case MediaType.CDROM:
+                case MediaType.GDROM: // TODO: Verify GD-ROM outputs this
+                    info.TracksAndWriteOffsets.Cuesheet = GenerateCuesheet(basePath + ".aif", basePath + ".cicm.xml") ?? "";
+                    break;
+            }
+
+            switch (system)
+            {
+
+            }
+        }
+
+        /// <summary>
         /// Get the list of commands that use a given flag
         /// </summary>
         /// <param name="flag">Flag value to get commands for</param>
@@ -1799,5 +1833,255 @@ namespace DICUI.Aaru
 
             return string.Empty;
         }
+
+        #region Information Extraction Methods
+
+        /// <summary>
+        /// Generate a cuesheet string based on CICM sidecar data
+        /// </summary>
+        /// <param name="aif">AIF output image</param>
+        /// <param name="cicmSidecar">CICM Sidecar data generated by Aaru</param>
+        /// <returns>String containing the cuesheet, null on error</returns>
+        private string GenerateCuesheet(string aif, string cicmSidecar)
+        {
+            // Note that this path only generates a cuesheet against a single BIN, not split
+
+            // If the files don't exist, we can't get info from them
+            if (!File.Exists(aif) || !File.Exists(cicmSidecar))
+                return null;
+
+            // If this is being run in Check, we can't run Aaru
+            if (string.IsNullOrEmpty(ExecutablePath))
+                return null;
+
+            var convertParams = new Parameters(null)
+            {
+                BaseCommand = Command.ImageConvert,
+
+                InputValue = aif,
+                OutputValue = aif + ".cue",
+            };
+
+            convertParams[Flag.XMLSidecar] = true;
+            convertParams.XMLSidecarValue = cicmSidecar;
+
+            // TODO: How do we get around this generation???
+            ExecuteInternalProgram(convertParams).ConfigureAwait(false).GetAwaiter().GetResult();
+
+            File.Delete(aif + ".bin");
+            return GetFullFile(aif + ".cue");
+        }
+
+        /// <summary>
+        /// Generate a CMP XML datfile string based on CICM sidecar data
+        /// </summary>
+        /// <param name="cicmSidecar">CICM Sidecar data generated by Aaru</param>
+        /// <returns>String containing the datfile, null on error</returns>
+        private string GenerateDatfile(string cicmSidecar)
+        {
+            // If the file doesn't exist, we can't get info from it
+            if (!File.Exists(cicmSidecar))
+                return null;
+
+            // Required variables
+            int totalTracks = 0;
+            string datfile = string.Empty;
+
+            // Open and read in the XML file
+            XmlReader xtr = XmlReader.Create(cicmSidecar, new XmlReaderSettings
+            {
+                CheckCharacters = false,
+                DtdProcessing = DtdProcessing.Ignore,
+                IgnoreComments = true,
+                IgnoreWhitespace = true,
+                ValidationFlags = XmlSchemaValidationFlags.None,
+                ValidationType = ValidationType.None,
+            });
+
+            // If the reader is null for some reason, we can't do anything
+            if (xtr == null)
+                return null;
+
+            // Only care about CICM sidecar files
+            xtr.MoveToContent();
+            if (xtr.Name != "CICMMetadata")
+                return null;
+
+            // Only care about OpticalDisc types
+            // TODO: Aaru - Support floppy images
+            xtr = xtr.ReadSubtree();
+            xtr.MoveToContent();
+            xtr.Read();
+            if (xtr.Name != "OpticalDisc")
+                return null;
+
+            // Get track count and all tracks now
+            xtr = xtr.ReadSubtree();
+            xtr.MoveToContent();
+            xtr.Read();
+            while (!xtr.EOF)
+            {
+                // We only want elements
+                if (xtr.NodeType != XmlNodeType.Element)
+                {
+                    xtr.Read();
+                    continue;
+                }
+
+                switch (xtr.Name)
+                {
+                    // Total track count
+                    case "Tracks":
+                        totalTracks = xtr.ReadElementContentAsInt();
+                        break;
+
+                    // Individual track data
+                    case "Track":
+                        XmlReader trackReader = xtr.ReadSubtree();
+                        if (trackReader == null)
+                            return null;
+
+                        int trackNumber = -1;
+                        long size = -1;
+                        string crc32 = string.Empty;
+                        string md5 = string.Empty;
+                        string sha1 = string.Empty;
+
+                        trackReader.MoveToContent();
+                        trackReader.Read();
+                        while (!trackReader.EOF)
+                        {
+                            // We only want elements
+                            if (trackReader.NodeType != XmlNodeType.Element)
+                            {
+                                trackReader.Read();
+                                continue;
+                            }
+
+                            switch (trackReader.Name)
+                            {
+                                // Track size
+                                case "Size":
+                                    size = trackReader.ReadElementContentAsLong();
+                                    break;
+
+                                // Track number
+                                case "Sequence":
+                                    XmlReader sequenceReader = trackReader.ReadSubtree();
+                                    if (sequenceReader == null)
+                                        return null;
+
+                                    sequenceReader.MoveToContent();
+                                    sequenceReader.Read();
+                                    while (!sequenceReader.EOF)
+                                    {
+                                        // We only want elements
+                                        if (sequenceReader.NodeType != XmlNodeType.Element)
+                                        {
+                                            sequenceReader.Read();
+                                            continue;
+                                        }
+
+                                        switch (sequenceReader.Name)
+                                        {
+                                            case "TrackNumber":
+                                                trackNumber = sequenceReader.ReadElementContentAsInt();
+                                                break;
+
+                                            default:
+                                                trackReader.Skip();
+                                                break;
+                                        }
+                                    }
+
+                                    // Skip the sequence now that we've processed it
+                                    trackReader.Skip();
+
+                                    break;
+
+                                // Checksums
+                                case "Checksums":
+                                    XmlReader checksumReader = trackReader.ReadSubtree();
+                                    if (checksumReader == null)
+                                        return null;
+
+                                    checksumReader.MoveToContent();
+                                    checksumReader.Read();
+                                    while (!checksumReader.EOF)
+                                    {
+                                        // We only want elements
+                                        if (checksumReader.NodeType != XmlNodeType.Element)
+                                        {
+                                            checksumReader.Read();
+                                            continue;
+                                        }
+
+                                        switch (checksumReader.Name)
+                                        {
+                                            case "Checksum":
+                                                string checksumType = checksumReader.GetAttribute("type");
+                                                string checksumValue = checksumReader.ReadElementContentAsString();
+                                                switch (checksumType)
+                                                {
+                                                    case "crc32":
+                                                        crc32 = checksumValue;
+                                                        break;
+                                                    case "md5":
+                                                        md5 = checksumValue;
+                                                        break;
+                                                    case "sha1":
+                                                        sha1 = checksumValue;
+                                                        break;
+                                                }
+
+                                                break;
+
+                                            default:
+                                                checksumReader.Skip();
+                                                break;
+                                        }
+                                    }
+
+                                    // Skip the checksums now that we've processed it
+                                    trackReader.Skip();
+
+                                    break;
+
+                                default:
+                                    trackReader.Skip();
+                                    break;
+                            }
+                        }
+
+                        // Build the track datfile data and append
+                        string trackName = Path.GetFileName(cicmSidecar).Replace(".cicm.xml", string.Empty);
+                        if (totalTracks == 1)
+                        {
+                            datfile += $"<rom name=\"{trackName}.bin\" size=\"{size}\" crc=\"{crc32}\" md5=\"{md5}\" sha1=\"{sha1}\" />\n";
+                        }
+                        else if (totalTracks > 1 && totalTracks < 10)
+                        {
+                            datfile += $"<rom name=\"{trackName} (Track {trackNumber}).bin\" size=\"{size}\" crc=\"{crc32}\" md5=\"{md5}\" sha1=\"{sha1}\" />\n";
+                        }
+                        else
+                        {
+                            datfile += $"<rom name=\"{trackName} (Track {trackNumber:2}).bin\" size=\"{size}\" crc=\"{crc32}\" md5=\"{md5}\" sha1=\"{sha1}\" />\n";
+                        }
+
+                        // Skip the track now that we've processed it
+                        xtr.Skip();
+
+                        break;
+
+                    default:
+                        xtr.Skip();
+                        break;
+                }
+            }
+
+            return datfile;
+        }
+
+        #endregion
     }
 }
